Restructured how the data is handled by adding Goals, Bios and Schedules in the Data folder.
I created 4 character Bios, 4 Schedules and 2 Goals (for Francis and Grace). The goals tell them to meet at roughly the same position at roughly the same time.
The game doesn't handle goals yet. In the future, it will generate tasks to carry out goals. For now, I wanted to get a sense of the data structures.

I also added a Zelda map and attached the camera to the player, just so we can get a sense of space and movement until the map system is in.